refactor(fsm): standardize init bitmasks and fix stack macro logic

+++ components/mqtt_service/mqtt_service.c
@@ -13,7 +13,7 @@
 // Defines the stack buffer for mqtt task
 
 #ifdef CONFIG_MQTT_STACK_SIZE
-    #define V_MQTT_STACK_BUFFER CONFIG_MQTT_STACK_SIZE
+    #define V_MQTT_STACK_BUFFER (uint16_t)CONFIG_MQTT_STACK_SIZE
 #else
     #define V_MQTT_STACK_BUFFER 2048
 #endif
